Bastion: Allowing the details container to scroll with a short window or long content.
