+++ engines/bastion/app/assets/bastion/incubator/stylesheets/tables.scss
@@ -109,9 +109,7 @@ $row-highlight: rgb(0, 162, 155) !default;
   width: 100%;
 }
 
-.table-scroll-wrapper {
-  position: fixed;
+.container-scroll-wrapper {
   overflow-x: hidden;
-  height: 100%;
-  width: 100%;
+  overflow-y: auto;
 }
