+++ engines/bastion/app/assets/bastion/systems/views/system-details.html
@@ -1,4 +1,4 @@
-<section class="system-details" ng-cloak>
+<section class="system-details" ng-cloak alch-container-scroll>
   <header class="details-header">
     <h2 class="fl">{{ "System \"%systemName\"" | i18n: {systemName: system.name} }}</h2>
