+++ engines/bastion/app/assets/bastion/systems/views/systems-table-collapsed.html
@@ -1,5 +1,5 @@
 <div alch-table="table" nutupane-table class="table-reduced" ng-class="{'table-mask': table.loadingMore}">
-  <div alch-table-scroll alch-infinite-scroll="table.nextPage()" class="table-reduced">
+  <div alch-container-scroll alch-infinite-scroll="table.nextPage()" class="table-full">
     <table class="table table-striped table-full">
       <thead>
       <tr alch-table-head row-select>
