+++ engines/bastion/app/assets/bastion/systems/views/systems-table-full.html
@@ -1,5 +1,5 @@
 <div alch-table="table" nutupane-table ng-class="{'table-mask': table.loadingMore}">
-  <div alch-table-scroll alch-infinite-scroll="table.nextPage()">
+  <div alch-container-scroll control-width="table" alch-infinite-scroll="table.nextPage()">
     <table class="table table-striped">
       <thead>
       <tr alch-table-head row-select>
